printer/pyproject.toml
[project]
name = "printer"
version = "1.1.0"
description = "Printer Skill for Openclaw"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "playwright>=1.58.0",
    "pycups>=2.0.4",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
]


[tool.uv]
dev-dependencies = []
